pyproject.toml
[project]
name = "massivegamearchive"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.1",
    "django-extensions>=4.1",
    "django-ninja>=1.4.1",
    "django-stubs>=5.2.0",
    "pillow>=11.2.1",
]
